requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=20.3.0=pypi_0
autobahn=20.12.3=pypi_0
automat=20.2.0=pypi_0
blas=1.0=mkl
ca-certificates=2021.1.19=haa95532_0
certifi=2020.12.5=py37haa95532_0
cffi=1.14.4=pypi_0
chardet=4.0.0=pypi_0
constantly=15.1.0=pypi_0
cryptography=3.3.1=pypi_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py37_0
dateparser=1.0.0=pypi_0
freetype=2.10.4=hd328e21_0
hyperlink=21.0.0=pypi_0
icu=58.2=ha925a31_3
idna=2.10=pypi_0
incremental=17.5.0=pypi_0
intel-openmp=2020.2=254
jpeg=9b=hb83a4c4_2
kiwisolver=1.3.0=py37hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py37hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py37h196d8e1_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mplfinance=0.12.7a5=pypi_0
ninja=1.10.2=py37h6d14046_0
numpy=1.19.2=py37hadc3359_0
numpy-base=1.19.2=py37ha3acd2a_0
olefile=0.46=py37_0
openssl=1.1.1i=h2bbff1b_0
pandas=1.2.1=py37hf11a4ad_0
pillow=8.1.0=py37h4fa10fc_0
pip=20.3.3=py37haa95532_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=pypi_0
pyhamcrest=2.0.2=pypi_0
pyopenssl=20.0.1=pypi_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h6538335_2
python=3.7.9=h60c2a47_0
python-binance=0.7.9=pypi_0
python-dateutil=2.8.1=py_0
pytorch=1.7.1=py3.7_cuda101_cudnn7_0
pytz=2020.5=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
regex=2020.11.13=pypi_0
requests=2.25.1=pypi_0
service-identity=18.1.0=pypi_0
setuptools=52.0.0=py37haa95532_0
sip=4.19.8=py37h6538335_0
six=1.15.0=py37haa95532_0
sqlite=3.33.0=h2a8f88b_0
tk=8.6.10=he774522_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu101
tornado=6.1=py37h2bbff1b_0
twisted=20.3.0=pypi_0
txaio=20.12.1=pypi_0
typing_extensions=3.7.4.3=pyh06a4308_0
tzlocal=2.1=pypi_0
ujson=4.0.2=pypi_0
urllib3=1.26.3=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zlib=1.2.11=h62dcd97_4
zope-interface=5.2.0=pypi_0
zstd=1.4.5=h04227a9_0
